pyproject.toml
# pyproject.toml
[tool.poetry]
name = "fraud-detection-project"
version = "0.1.0"
description = "A production-grade MLOps pipeline for fraud detection."
authors = ["MLOps Engineer <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
xgboost = "^2.0.3"
lightgbm = "^4.4.0"
mlflow = "^2.13.2"
dvc = {extras = ["s3"], version = "^3.50.1"}
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
boto3 = "^1.34.119"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
jupyter = "^1.0.0"
joblib = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.5.0"